UWAGA !!! Refactor obiektow zakonczony

+++ client/src/main/java/com/crashcourse/restclient/datatype/PrzyjecieZamowieniaTO.java
@@ -5,16 +5,16 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import java.util.Date;
 
 public class PrzyjecieZamowieniaTO {
-    private int ID;
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
+    private Long ID;
+    @JsonFormat(shape=JsonFormat.Shape.STRING, pattern="yyyy-MM-dd")
     private Date data;
     private ZamowienieDostawyTO zamowienie;
 
-    public int getID() {
+    public Long getID() {
         return ID;
     }
 
-    public void setID(int ID) {
+    public void setID(Long ID) {
         this.ID = ID;
     }
